backend/requirements.txt
Django
djangorestframework
gunicorn
paho-mqtt
celery
redis